css/css-position/sticky/sticky-text-mouse-drag.html

<!doctype html>
<title>Sticky positioned element should block propagate scrolling after mouse selection</title>
<link rel="author" title="Seokho Song"  href="mailto:0xdevssh@gmail.com">
<link rel="help" href="https://crbug.com/1178622">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<div id="scrollEl" style="height:100px; overflow-y: auto; scroll-padding:20px 20px 20px 20px;">
    <input id="stickyEl" style="position:sticky;top:0" value="This iaasdks a input text"/>
    <div>1</div>
    <div>2</div>
    <div>3</div>
    <div>4</div>
    <div>5</div>
    <div>6</div>
    <div>7</div>
    <div>8</div>
    <div>9</div>
    <div>10</div>
    <div>11</div>
    <div>12</div>
    <div>13</div>
    <div>14</div>
    <div>15</div>
    <div>16</div>
    <div>17</div>
</div>
</div>
<script>

async_test(t => {
    var scrollEl = document.getElementById("scrollEl");
    var stickyEl = document.getElementById("stickyEl");
    var rect = stickyEl.getBoundingClientRect()

    scrollEl.scrollTo(0, scrollEl.scrollHeight);
    const old_scroll_pos = scrollEl.scrollTop

    let actionsPromise = new test_driver.Actions()
    .pointerMove(rect.left, rect.top + 5)
    .pointerDown()
    .pointerMove(rect.left + 100, rect.top + 5)
    .pointerUp()
    actionsPromise.send().then(
        t.step_func(()=>{
            assert_equals(scrollEl.scrollTop, old_scroll_pos,
            "should not scroll up after mouse drag text");
            t.done()
        }
    ))
})
</script>
